+++ objc/Source/Genomics/TTIOBulkV2Blobs.m
@@ -1,0 +1,11 @@
+/*
+ * TTIOBulkV2Blobs.m
+ * TTI-O Objective-C Implementation
+ *
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ */
+#import "TTIOBulkV2Blobs.h"
+
+@implementation TTIOBulkV2Blobs
+
+@end